Add support for credits in the about dialog of the generated project. Automatically export/import credits when objects are exported and imported. Automatically copy dependent DLL files for imported code objects if they exist externally. Update test project. Add partial music library.

+++ SGDK2IDE/Template/LayerBase.cs
@@ -259,8 +259,8 @@ public abstract class LayerBase : System.Collections.IEnumerable
                    (CurFrame.priority < 0))
             {
                spr.Transform = Matrix.Multiply(CurFrame.frame.Transform, Matrix.Translation(
-                  CurFrame.x + CurrentPosition.X + ViewRect.X,
-                  CurFrame.y + CurrentPosition.Y + ViewRect.Y, 0));
+                  (float)CurFrame.x + CurrentPosition.X + ViewRect.X,
+                  (float)CurFrame.y + CurrentPosition.Y + ViewRect.Y, 0));
                spr.Draw(CurFrame.frame.GraphicSheetTexture.Texture, CurFrame.frame.SourceRect,
                   Vector3.Empty, Vector3.Empty, CurFrame.color);
                if (!Injected.MoveNext())
@@ -289,8 +289,8 @@ public abstract class LayerBase : System.Collections.IEnumerable
       {
          CurFrame = (InjectedFrame)Injected.Current;
          spr.Transform = Matrix.Multiply(CurFrame.frame.Transform, Matrix.Translation(
-            CurFrame.x + CurrentPosition.X + ViewRect.X,
-            CurFrame.y + CurrentPosition.Y + ViewRect.Y, 0));
+            (float)CurFrame.x + CurrentPosition.X + ViewRect.X,
+            (float)CurFrame.y + CurrentPosition.Y + ViewRect.Y, 0));
          spr.Draw(CurFrame.frame.GraphicSheetTexture.Texture, CurFrame.frame.SourceRect,
             Vector3.Empty, Vector3.Empty, CurFrame.color);
          if (!Injected.MoveNext())
